fetch_html_screenshot_with_selenium/requirements.txt
selenium==4.28.1
boto3
Pillow